Refactor: Extract `detect::windows` in crate `detect-targets`
