+++ crates/detect-targets/src/detect.rs
@@ -77,22 +77,4 @@ mod linux;
 mod macos;
 
 #[cfg(target_os = "windows")]
-mod windows {
-    use crate::TARGET;
-    use guess_host_triple::guess_host_triple;
-
-    pub(super) fn detect_alternative_targets(target: &str) -> Option<String> {
-        let (prefix, abi) = target.rsplit_once('-').expect("Invalid target triple");
-
-        // detect abi in ["gnu", "gnullvm", ...]
-        (abi != "msvc").then(|| format!("{prefix}-msvc"))
-    }
-
-    pub(super) fn detect_targets_windows() -> Vec<String> {
-        let mut targets = vec![guess_host_triple().unwrap_or(TARGET).to_string()];
-
-        targets.extend(detect_alternative_targets(&targets[0]));
-
-        targets
-    }
-}
+mod windows;
